ref: Use Metrics for jingle stats.
